% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_depths}
\alias{barplot_depths}
\title{Create a Bar Plot of Abundances from Phyloseq Data with Grouping and Fill Options}
\usage{
barplot_depths(
  ps,
  group = "Sample",
  fill = NULL,
  position = "stack",
  wrap = NULL,
  ...
)
}
\arguments{
\item{ps}{A phyloseq object or a list of phyloseq objects. If a list is provided, all objects are combined.}

\item{group}{The name of the column to be used for grouping on the x-axis. Defaults to "Sample".}

\item{fill}{(Optional) The name of the column to apply a fill color to bars. NULL by default, which results in no fill.}

\item{position}{Determines how to position bars. "stack" stacks bars on top of each other, "dodge" places them side by side. Defaults to "stack".}

\item{wrap}{(Optional) The name of a column to be used for facetting the plot into separate panels.}

\item{...}{Additional arguments passed on to \code{geom_col} from \code{ggplot2}.}
}
\value{
A \code{ggplot} object representing the bar plot. Can be printed or modified further.
}
\description{
Create a Bar Plot of Abundances from Phyloseq Data with Grouping and Fill Options
}
\examples{
data("GlobalPatterns", package = "phyloseq")
barplot_depths(GlobalPatterns, group = "SampleType", fill = "Phylum")
barplot_depth(ps_list, group = "group", fill = "Gene")
barplot_depth(ps_list, group = "group", fill = "Phylum")
barplot_depth(ps_list, group = "Gene", fill = "Phylum")
barplot_depth(ps_list, group = c("Gene", "group"), fill = "Phylum") + facet_wrap(~group, scales = "free")


}
