% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot_depths}
\alias{boxplot_depths}
\title{Generate a Boxplot with Significance Indicators}
\usage{
boxplot_depths(
  ps,
  fill = NULL,
  x = fill,
  wrap = NULL,
  signif = TRUE,
  test = "wilcox.test",
  ...
)
}
\arguments{
\item{ps}{A phyloseq object or a list of phyloseq objects. The
function melts the data for plotting.}

\item{fill}{The name of the variable to use for filling boxes. Can be
set to NULL. Defaults to NULL.}

\item{x}{The name of the variable to use for the x-axis. Inherits the
value of \code{fill} by default.}

\item{wrap}{An optional string specifying a variable name to wrap the
plot into multiple panels. Not supported alongside \code{signif}.}

\item{signif}{Logical, indicating whether to perform and display
significance testing between groups. Defaults to TRUE.}

\item{test}{The name of the statistical test function to use as a
character string. Defaults to "wilcox.test".}

\item{...}{Additional arguments passed to \code{geom_boxplot}.}
}
\value{
A ggplot object representing the boxplot with or without
significance indicators.
}
\description{
Generate a Boxplot with Significance Indicators
}
\details{
This function creates a boxplot of abundance data from a
given phyloseq object or a list of phyloseq objects. When \code{signif} is
TRUE, it automatically calculates and displays significance levels
between the comparisons of boxplots based on the specified statistical
test. \code{wrap} parameter allows creating multiple panels but is not
compatible with \code{signif} parameter. If using \code{signif}, consider using
\code{patchwork} for combining multiple plots.
}
\note{
Using \code{wrap} together with \code{signif} will result in an error
message as they are not compatible. Use \code{patchwork} package for layout
management in such cases.
}
\examples{
# Assuming `ps` is a phyloseq object with proper setup:
boxplot_depths(ps, fill = "SampleType", x = "Condition", signif = TRUE)
# For multiple phyloseq objects in a list and using a Wilcoxon test:
boxplot_depths(list(ps1, ps2), fill = "SampleType", x = "Condition", test = "wilcox.test", signif = TRUE)

}
