% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{build_hmm}
\alias{build_hmm}
\title{Build a Hidden Markov Model (HMM) from a sequence alignment}
\usage{
build_hmm(aln)
}
\arguments{
\item{aln}{A sequence alignment object. The alignment object should have an optional "name" attribute, which will be used as the name of the HMM.}
}
\value{
The path to the file containing the generated HMM.
}
\description{
This function takes a sequence alignment, converts it to the Stockholm format, and then uses the \code{hmmbuild} command
from the HMMER suite to build a Hidden Markov Model (HMM).
}
\note{
This function requires the external programs \code{esl-reformat} and \code{hmmbuild} from the HMMER suite to be
installed and accessible in the system's PATH.
}
\examples{
# Assuming `my_alignment` is an alignment object
hmm_file <- build_hmm(my_alignment)
# hmm_file now contains the path to the HMM file

}
\references{
Eddy, S.R. "HMMER: biosequence analysis using profile hidden Markov models."
http://hmmer.org/
}
