% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{get_contigs_hits_depths_from_hmm}
\alias{get_contigs_hits_depths_from_hmm}
\title{Get Contigs Hits and Depths from HMM Searches}
\usage{
get_contigs_hits_depths_from_hmm(
  queries_hmm,
  control_hmm,
  depths_files,
  sub_pattern,
  sub_replacement,
  dbs,
  linker = contigs_linker,
  parallel_processes = 1,
  cpus_per_process = 1,
  incE = 1e-06,
  hmmer_path = "",
  verbose = FALSE
)
}
\arguments{
\item{queries_hmm}{Character vector specifying the file paths to the
query HMM profiles.}

\item{control_hmm}{Character string specifying the file path to the
control HMM profile.}

\item{depths_files}{Character vector with paths to the depth files.}

\item{sub_pattern}{Character string representing the pattern to
identify contig names within depth files.}

\item{sub_replacement}{Character string with the replacement for the
sub_pattern in contig names.}

\item{dbs}{Character vector specifying the paths to the databases to
search against.}

\item{linker}{Function that links hits to contigs. If not supplied, a
default contigs_linker function will be used.}

\item{incE}{Numeric, inclusion E-value threshold for HMM searches.
Defaults to 1e-6.}

\item{hmmer_path}{Character string specifying the path to the HMMER
binaries. If empty, system's PATH will be used.}

\item{verbose}{Logical indicating whether to print detailed progress
messages. Defaults to FALSE.}

\item{cpu}{Integer specifying the number of CPUs to use for HMM
searches. Defaults to 1.}
}
\value{
A list where each element corresponds to the results from
each query HMM, containing detailed hits and depths information for
contigs.
}
\description{
This function performs HMM searches for query and control sequences
against a set of databases, retrieves hits, and calculates depths for
contigs.
}
\examples{
# Assuming appropriate files and HMM profiles exist:
result <- get_contigs_hits_depths_from_hmm(
  queries_hmm = c("/path/to/query1.hmm", "/path/to/query2.hmm"),
  control_hmm = "/path/to/control.hmm",
  depths_files = c("/path/to/depth1.tsv", "/path/to/depth2.tsv"),
  sub_pattern = "_[0-9]+$",
  sub_replacement = "",
  dbs = c("/path/to/db1", "/path/to/db2"),
  linker = custom_linker_function,
  cpu = 4, incE = 1e-10,
  hmmer_path = "/opt/hmmer/bin/",
  verbose = TRUE)

}
