% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{get_hits_depths_from_kos}
\alias{get_hits_depths_from_kos}
\title{Retrieve depths of hits from KEGG Orthology entries across samples}
\usage{
get_hits_depths_from_kos(
  kos,
  ps,
  dbs,
  control_tblout,
  linker,
  msa_method = "Muscle",
  cpu = 1,
  incE = 1e-06,
  taxrank = NULL,
  ...
)
}
\arguments{
\item{kos}{A character vector of KEGG Orthology (KO) identifiers.}

\item{ps}{A phyloseq object containing the samples data.}

\item{dbs}{A character vector indicating the databases to be used.}

\item{control_tblout}{A tblout formatted file with the control genes' hits.}

\item{linker}{Function that links OTU identifiers between tblout tables and
phyloseq object. The function must take two arguments: SeqFile and Target.}

\item{msa_method}{The method used for multiple sequence alignment. "Muscle"
by default.}

\item{cpu}{The number of CPUs to be used for computation. Integer value.
1 by default.}

\item{incE}{The inclusion E-value threshold for considering a hit
significant. Numeric value. 1e-6 by default.}

\item{taxrank}{Optional; character specifying the taxonomic rank at which
to aggregate hits. Valid options depend on the taxonomic ranks present in
the \code{ps} object. If not specified, no aggregation is performed.}
}
\value{
A list where each element corresponds to a KO identifier
inputted and contains the computed depths of hits for that KO in the
sample set.
}
\description{
This function takes KEGG Orthology (KO) identifiers, a phyloseq
object, a set of databases, a control tblout formatted file, and
computes the depths of hits related to the input KOs in the provided
sample set.
}
\examples{
kos <- c("K00001", "K00002")
ps <- phyloseq_object # a pre-loaded or created phyloseq object
dbs <- c("db1", "db2")
control_tblout <- "control_hits.tblout"
linker <- read.csv("linker_table.csv")
hit_depths <- get_hits_depths_from_kos(kos, ps, dbs, control_tblout, linker)

}
