% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{get_kegg_kos_from_module}
\alias{get_kegg_kos_from_module}
\title{Retrieve KEGG Orthology IDs (KOs) associated with a KEGG module}
\usage{
get_kegg_kos_from_module(module)
}
\arguments{
\item{module}{A character string specifying the KEGG module ID (e.g., "M00001").
Only one module ID should be provided.}
}
\value{
A character vector of KEGG Orthology IDs (KOs) associated with the
specified module. If multiple KOs are associated, they are returned
as separate elements in the vector.
}
\description{
This function fetches KEGG Orthology IDs (KOs) that are associated with a
specified KEGG module. It queries the KEGG database for the module, extracts
the KOs, and returns them as a vector.
}
\note{
: The function relies on the \code{KEGGREST} package for fetching data
from the KEGG database. Ensure that this package is installed.
into your R session.
}
\examples{
kegg_kos <- get_kegg_kos_from_module("M00001")
print(kegg_kos)

}
