% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{get_kegg_msa}
\alias{get_kegg_msa}
\title{Retrieve and Align KEGG Gene Sequences}
\usage{
get_kegg_msa(ko, nmax = Inf, method = "Muscle", ...)
}
\arguments{
\item{ko}{A character string specifying the KEGG Orthology (KO) identifier.}

\item{nmax}{The maximum number of genes to include in the alignment (Default: no limit).}

\item{...}{Additional arguments passed to the \code{msa} function for performing the multiple sequence alignment.}
}
\value{
An object of class \code{msa} representing the multiple sequence alignment of the retrieved sequences. The name of the KO identifier (or, if NULL, the KO identifier itself) is stored as an attribute named \code{"name"} of the returned object.
}
\description{
This function retrieves amino acid sequences for genes associated with a given KEGG Orthology (KO) identifier and performs a multiple sequence alignment (MSA) on them.
}
\examples{
\dontrun{
  aln <- get_kegg_msa("K02533")
  plot(aln)
}

}
