% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverted_names}
\alias{inverted_names}
\title{Invert Names in a Named Vector}
\usage{
inverted_names(v)
}
\arguments{
\item{v}{A named vector.}
}
\value{
A named vector with its names and values swapped.
}
\description{
Swaps the names and values in a named vector to create a new vector
with values as names and names as values.
}
\examples{
my_vector <- c("name1" = "Alice", "name2" = "Bob", "name3" = "Charlie")
inverted_names(my_vector)
# Returns: Alice = "name1", Bob = "name2", Charlie = "name3"

}
