% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{keggmodule_plot}
\alias{keggmodule_plot}
\title{Plot Enzyme Depths for a Given KEGG Module and Depths List}
\usage{
keggmodule_plot(kegg_module, depths_list, wrap = NULL, plot = TRUE)
}
\arguments{
\item{kegg_module}{The module name as a string (e.g. "M00001") or
a data.table containing the KEGG pathway graph data (derived from
get_kegg_reactions_from_module()).}

\item{depths_list}{A list of phyloseq objects representing the depths.}

\item{wrap}{An optional column name as string for faceting.
If \code{NULL}, no wrapping is performed.}

\item{plot}{If FALSE, do not plot anything, just return the raw table.}
}
\value{
A ggraph plot of the specified KEGG module, annotated with
enzyme depths. if \code{plot} is FALSE, a tidygraph object with the depths
information will be returned instead of the plot.
}
\description{
This function takes a KEGG pathway graph and a list of
depths data, aggregates the depths information, and creates an igraph
object representing the KEGG pathway with aggregated node abundances.
}
