% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_motus_tax_table}
\alias{make_motus_tax_table}
\title{Generate a basic colData file from a mOTUs file}
\usage{
make_motus_tax_table(path, level = "mOTU")
}
\arguments{
\item{path}{Path to a mOTUs file}

\item{level}{Taxonomic level of the mOTUs in the file}
}
\description{
Generate a basic colData file from a mOTUs file
}
