% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_motus_tax_table_from_gtdb}
\alias{make_motus_tax_table_from_gtdb}
\title{Create taxonomic table for mOTUs using GTDB identifiers}
\usage{
make_motus_tax_table_from_gtdb(motus_table, gtdb_map)
}
\arguments{
\item{motus_table}{The user's mOTUs table with rownames as mOTUs identifiers.}

\item{gtdb_map}{Path to the GTDB mapping file.}
}
\value{
A \code{tax_table} object containing the GTDB taxonomy for the input mOTUs.
}
\description{
This function takes a user's mOTUs table and a mapping file between mOTUs and GTDB
identifiers to create a new taxonomic table with GTDB taxonomy.
}
\note{
Please ensure the GTDB mapping file is available at the specified path. For mOTUs3,
the file can be downloaded from https://sunagawalab.ethz.ch/share/MOTU_GTDB/mOTUs_3.0.0_GTDB_tax.tsv.
If the file is not found, the function will stop and prompt to download or specify the mapping file.
}
\examples{
# Assuming you've already downloaded the GTDB mapping file:
gtdb_map_path <- "path/to/mOTUs_3.0.0_GTDB_tax.tsv"
motus_table <- read.csv("your_motus_table.csv", row.names = 1)
gtdb_tax_table <- make_motus_tax_table_from_gtdb(motus_table, gtdb_map_path)

}
