% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merged_hits_to_otu_table}
\alias{merged_hits_to_otu_table}
\title{Merged Hits to OTU Table}
\usage{
merged_hits_to_otu_table(mer, taxa_are_rows)
}
\arguments{
\item{mer}{A \code{data.table} with matched pairs of columns for each
taxon, each pair consisting of a ".x" and ".y" column, and a column
named "ID" that contains the taxa names or IDs.}

\item{taxa_are_rows}{A logical parameter indicating whether taxa
should be represented by rows (TRUE) or by columns (FALSE) in the
final OTU table.}
}
\value{
Returns an \code{otu_table} object with ratios of hits, with
NaNs and Infs set to NA. If \code{taxa_are_rows} is TRUE, taxa will be
represented by rows; otherwise, they will be represented by columns.
}
\description{
This function takes a merged hits data.table (with queries and
controls), calculates the depths ratio, and returns an OTU
(Operational Taxonomic Unit) table, optionally transposing the table
based on taxa orientation.
It operates on the result of get_hits_depths() when the \code{phyloseq}
argument is FALSE.
}
\details{
This function cleans the resulting matrix by replacing NaNs and Infs
with NAs.
}
\examples{
# Assuming `dt_hits` is a data.table with matched pairs of hit columns (e.g., "gene1.x", "gene1.y")
# and an "ID" column with taxa names.
otu_table <- merged_hits_to_otu_table(dt_hits, TRUE)

}
