% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_hmmer_tblout}
\alias{read_hmmer_tblout}
\title{Read HMMER tblout files}
\usage{
read_hmmer_tblout(file)
}
\arguments{
\item{file}{A string. The path to the HMMER tblout file to be read.}
}
\value{
A data.table object containing the parsed HMMER tblout data. Columns
include Target, Accession_target, Query, Accession_query, Full_evalue,
Full_score, Full_bias, Best_Evalue, Best_score, Best_bias, exp, reg, clu,
ov, env, dom, rep, inc, and Description.
}
\description{
Parses the tblout output format of HMMER tools (e.g., hmmscan, hmmsearch) and
creates a data table with relevant columns for target and query sequences,
including e-values, scores, and additional metrics, plus the description.
}
\examples{
read_hmmer_tblout("path/to/your/hmmer_output.tblout")

}
