% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{search_hmm}
\alias{search_hmm}
\title{Perform HMMER search across multiple databases}
\usage{
search_hmm(hmm, dbs, cpu = 1, incE = 1e-06)
}
\arguments{
\item{hmm}{A string specifying the path to the HMM file to be searched with.}

\item{dbs}{A character vector, where each element is a path to a database file to be searched.}

\item{cpu}{An integer indicating the number of CPUs to be used for the search. Defaults to 1.}

\item{incE}{A numeric value specifying the inclusion E-value threshold. Defaults to 1e-6.}
}
\value{
A data frame containing the combined results of the HMM searches on all databases.
Each row represents one HMMER hit, and the data frame includes a column 'SeqFile' indicating
the database file from which each hit originates.
}
\description{
This function executes HMMER search on a given Hidden Markov Model (HMM) file across
multiple database files (dbs). It allows the setting of the number of CPUs to be used
and the inclusion E-value threshold. The results from searching each database are combined
into a single data table.
}
\examples{
hmm_file <- "your_hmm_file.hmm"
db_files <- c("database1.fasta", "database2.fasta")
search_results <- search_hmm(hmm_file, db_files, cpu = 2, incE = 1e-5)

}
