% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{set_group_order}
\alias{set_group_order}
\title{Set custom order for groups in phyloseq object}
\usage{
set_group_order(ps, group, levels)
}
\arguments{
\item{ps}{A phyloseq object containing sample data.}

\item{group}{The name of the group (column in the sample data) to set the order for. If the group is "Sample", the function treats it as a special case.}

\item{levels}{A vector of values defining the new order of the group.}
}
\value{
A modified phyloseq object with the new order set for the specified group.
}
\description{
This function sets a custom ordering for a specified group within the sample data of a phyloseq object. The new ordering is stored in a new column within the sample data, distinguishing between a special case for the "Sample" group and other groups.
}
\examples{
# Given a phyloseq object `ps`, set a new order for the "Sample" group
ps_new <- set_group_order(ps, "Sample", c("Sample3", "Sample1", "Sample2"))

# For a group "Treatment", set a new order
ps_new <- set_group_order(ps, "Treatment", c("Control", "Treatment1", "Treatment2"))

}
