% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shorten_species_name}
\alias{shorten_species_name}
\title{Shorten Common Name to Species Initialism}
\usage{
shorten_species_name(common_name)
}
\arguments{
\item{common_name}{A string containing the common name of a species.}
}
\value{
A string with the abbreviated species name.
}
\description{
This function takes a common species name and shortens it into an initialism
format consisting of the first letter of the first word followed by a period
and the full second word. Leading and trailing whitespaces are trimmed.
Common long words like "Candidatus" and "species incertae sedis" are abbreviated.
Useful for reducing the legend clutter when plotting.
}
\examples{
shorten_species_name("Canis lupus")
# [1] "C. lupus"
shorten_species_name("Lynx")
# Lynx
shorten_species_name("Candidatus Mycoplasma hyopneumoniae")
# Ca. M. hyopneumoniae
shorten_species_name("Proteobacteria species incertae sedis")
# Proteobacteria sp. inc. s.
}
