% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tblout_from_afa}
\alias{tblout_from_afa}
\title{Perform HMM-based search from an alignment}
\usage{
tblout_from_afa(afa, dbs, cpu = 1, incE = 1e-06)
}
\arguments{
\item{afa}{Aligned fasta sequences in string format or an object that
can be interpreted as such.}

\item{cpu}{The number of CPU cores to use for the search process.
Defaults to 1 to ensure compatibility with all systems.}

\item{incE}{The inclusion threshold E-value for reported hits in the
search. Lower values are more stringent. Defaults to 1e-6.}

\item{...}{Additional arguments to be passed to the search function.}
}
\value{
Returns the path to the tblout results.
}
\description{
This function takes aligned fasta sequences (AFA) as an input, constructs
a Hidden Markov Model (HMM), and uses it to search in designated databases.
}
\examples{
# Example assuming 'sequences.afa' is your aligned fasta file
afa_content <- readLines("sequences.afa")
search_results <- tblout_from_afa(afa_content)
print(search_results)

}
