% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tblout_from_faa}
\alias{tblout_from_faa}
\title{Perform a Multiple Sequence Alignment and Find Hits in Databases}
\usage{
tblout_from_faa(faa, dbs, method = "Muscle", cpu = 1, incE = 1e-06, ...)
}
\arguments{
\item{faa}{Character. The path to the fasta file containing amino acid sequences.}

\item{dbs}{Character or Character vector. Specifies the databases against which the
sequences will be searched.}

\item{method}{Character. The default method for sequence alignment is "Muscle".
Other methods supported by \code{msa} function can be used.}

\item{cpu}{Integer. The number of CPU threads to use for the search. The default is 1.}

\item{incE}{Numeric. The inclusion threshold E-value for considering a database hit
significant. Defaults to 1e-6.}

\item{...}{Additional arguments passed to the \code{msa} function.}
}
\value{
Returns the result of searching the aligned sequences against specified
databases. The format of the returned object depends on the implementation of
\code{tblout_from_afa}.
}
\description{
This function first reads a fasta file containing amino acid sequences, performs
a multiple sequence alignment using the specified method, and then searches
for hits in databases using the aligned sequences. The results are returned
in a specific format.
}
\examples{
\dontrun{
tblout_from_faa("sequences.faa", dbs=c("db1", "db2"), method="ClustalW")
}

}
