% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tblout_from_ko}
\alias{tblout_from_ko}
\title{Perform HMM Search on a KEGG Orthology (KO) Alignment}
\usage{
tblout_from_ko(ko, dbs, method = "Muscle", cpu = 1, incE = 1e-06, ...)
}
\arguments{
\item{ko}{A character string representing the KEGG Orthology (KO) identifier.}

\item{method}{Character. The method used for performing the MSA. Defaults to "Muscle".}

\item{cpu}{Integer. The number of CPUs to use for the search. Defaults to 1.}

\item{incE}{Numeric. The inclusion threshold for the HMM search. Defaults to 1e-6.}

\item{...}{Additional arguments passed to kegg_kegg_msa().}
}
\value{
The result of the HMM search.
}
\description{
Given a KO identifier, this function retrieves the corresponding protein sequences and builds a Multiple Sequence Alignment (MSA), then builds a Hidden Markov Model (HMM) based on the MSA, and then searches
the HMM against a set of sequence databases. The function allows adjusting the search method,
computational resources, and inclusion threshold.
}
\examples{
tblout_from_ko("K00001")
}
