% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_motus}
\alias{import_motus}
\title{Import mOTUs data}
\usage{
import_motus(path, make_tax_table = F)
}
\arguments{
\item{path}{character string specifying the location of the file containing the mOTUs data.}

\item{make_tax_table}{logical indicating whether a taxonomic table should also be created and attached
to the phyloseq object. Defaults to FALSE.}
}
\value{
The imported motus data. If make_tax_table is TRUE, a tax table is also returned.

Returns a phyloseq object constructed from the OTU data, with an optional taxonomic table.
}
\description{
Keep only the motus ID, not the tax name, which should be provided separately as colData.
}
\details{
This function is used to import mOTUs data from a specified path.
Optionally, it can also create a default tax table from the mOTUs names.
}
\examples{
motus_data <- import_motus("path/to/your/motus_file.csv", make_tax_table = TRUE)
}
